chore: move fetch functions to a separate util file
